+++ app/Http/Controllers/BookController.php
@@ -20,30 +20,35 @@ class BookController extends Controller
 
     function create()
     {
+        $this->authorize('create');
         $authors = \App\Author::all();
         return view('books.create', compact('authors'));
     }
 
     function store()
     {
+        $this->authorize('create');
         Book::create($this->getValidBook());
         return redirect(route('books.index'));
     }
 
     function edit(Book $book)
     {
+        $this->authorize('update');
         $authors = \App\Author::all();
         return view('books.edit', compact('book', 'authors'));
     }
 
     function update(Book $book)
     {
+        $this->authorize('update');
         $book->update($this->getValidBook());
         return redirect(route('books.show', $book));
     }
 
     function delete(Book $book)
     {
+        $this->authorize('update');
         $book->delete();
         return redirect(route('books.index'));
     }
